+++ dev-tools/config.json
@@ -138,7 +138,7 @@
     "DDEX_KEY": "49d5e13d355709b615b7cce7369174fb240b6b39",
     "DDEX_SECRET": "2b2c2b90d9a489234ae629a5284de84fb0633306257f17667aaebf2345d92152",
     "SESSION_SECRET": "something random",
-    "DDEX_ADMIN_ALLOWLIST": "127559427",
+    "DDEX_ADMIN_ALLOWLIST": "127559427,555164012",
     "NODE_ENV": "stage"
   }
 }
