+++ packages/discovery-provider/src/tasks/entity_manager/utils.py
@@ -1,6 +1,6 @@
 import json
 import os
-from datetime import datetime
+from datetime import datetime, timezone
 from enum import Enum
 from typing import Dict, List, Literal, Set, Tuple, TypedDict, Union
 
@@ -512,3 +512,30 @@ def is_ddex_signer(signer):
             address.lower() for address in ddex_apps.split(",")
         )
     return False
+
+
+def parse_release_date(release_date_str):
+    # try various time formats
+    if not release_date_str:
+        return None
+
+    try:
+        return datetime.strptime(
+            release_date_str, "%a %b %d %Y %H:%M:%S GMT%z"
+        ).astimezone(timezone.utc)
+    except ValueError:
+        pass
+
+    try:
+        return datetime.strptime(release_date_str, "%Y-%m-%dT%H:%M:%S.%fZ").astimezone(
+            timezone.utc
+        )
+    except ValueError:
+        pass
+
+    try:
+        return datetime.fromtimestamp(int(release_date_str)).astimezone(timezone.utc)
+    except (ValueError, TypeError):
+        pass
+
+    return None
